% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_key.R
\name{draw_key_plottwist2}
\alias{draw_key_plottwist2}
\alias{draw_key_patch}
\alias{draw_key_symbol}
\title{More key glyphs for legends}
\usage{
draw_key_patch(data, params, size)

draw_key_symbol(data, params, size)
}
\arguments{
\item{data}{A single row data frame containing the scaled aesthetics to
display in this key}

\item{params}{A list of additional parameters supplied to the geom.}

\item{size}{Width and height of key in mm.}
}
\description{
An expansion of the draw_key_x functions from ggplot2 to include the new
geoms included in plottwist2.
}
