% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_rotate.R
\name{symbol_flip}
\alias{symbol_flip}
\title{Flip a symbol}
\usage{
symbol_flip(recipe, axis = "x")
}
\arguments{
\item{recipe}{A string generated by \code{symbol_recipe()} or found in \code{symbol_lookup}.}

\item{axis}{The axis for the symbol to be mirrored along. \code{x}, \code{y}, or \code{xy} for mirroring along both axes.}
}
\value{
A string in the same format as a \code{symbol_recipe()} output, for use with \code{geom_symbol()}
}
\description{
Mirror a symbol for use with geom_symbol.
}
