% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_recipe.R
\name{symbol_recipe}
\alias{symbol_recipe}
\title{Define new symbol}
\usage{
symbol_recipe(x, y, id = NA)
}
\arguments{
\item{x}{A numeric vector specifying x values.}

\item{y}{A numeric vector specifying y values.}

\item{id}{A numeric vector used to separate locations in x and y into
multiple lines. All locations with the same \code{id} belong to the same line.}
}
\value{
An string which can be passed to \code{geom_symbol()} or
\code{geom_symbolswarm()} to be interpreted.
}
\description{
Supply coordinates for a symbol to be plotted with \code{geom_symbol()} or
\code{geom_symbolswarm()}.
}
