% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/symbol_rotate.R
\name{symbol_rotate}
\alias{symbol_rotate}
\title{Rotate a symbol}
\usage{
symbol_rotate(recipe, degrees)
}
\arguments{
\item{recipe}{A string generated by \code{symbol_recipe()} or found in \code{symbol_lookup}.}

\item{degrees}{The degree of clockwise rotation desired, in degrees rather than radians.}
}
\value{
A string in the same format as a \code{symbol_recipe()} output, for use with \code{geom_symbol()}
}
\description{
Rotate a symbol for use with geom_symbol.
}
