% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_simple.R
\name{theme_simple}
\alias{theme_simple}
\title{Simple Theme for ggplot2}
\usage{
theme_simple()
}
\value{
A 'theme' class object which can be added to a plot.
}
\description{
ggplot2 theme which simplifies everything to black and white lines of uniform thickness. Text is converted to element_markdown() from the ggtext package, allowing for use of italics etc.
}
\examples{
p <- ggplot2::ggplot(data = mtcars, ggplot2::aes(x = cyl, y = mpg)) +
    ggplot2::geom_point() +
    theme_simple()

}
